third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-palette-expected.txt

This is a testharness.js-based test.
[FAIL] 'font-palette' does not support 'Augusta'
  assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object Undefined]"
Harness: the test ran to completion.
